session: move scoped summary helper to internal package
