+++ session/internal/summary/summarizer.go
@@ -11,29 +11,12 @@ package summary
 import (
 	"context"
 
-	"trpc.group/trpc-go/trpc-agent-go/model"
 	"trpc.group/trpc-go/trpc-agent-go/session"
+	"trpc.group/trpc-go/trpc-agent-go/session/summary"
 )
 
-// SessionSummarizer is the internal summary interface used by session services.
-// It intentionally mirrors the public session/summary interface so internal
-// helpers can avoid importing the public package and creating dependency cycles.
-type SessionSummarizer interface {
-	ShouldSummarize(sess *session.Session) bool
-	Summarize(ctx context.Context, sess *session.Session) (string, error)
-	SetPrompt(prompt string)
-	SetModel(m model.Model)
-	Metadata() map[string]any
-}
-
-// ContextAwareSummarizer is the context-aware extension of SessionSummarizer.
-type ContextAwareSummarizer interface {
-	SessionSummarizer
-	ShouldSummarizeWithContext(context.Context, *session.Session) bool
-}
-
 // HasSummarizer reports whether summary generation is configured.
-func HasSummarizer(summarizer SessionSummarizer) bool {
+func HasSummarizer(summarizer summary.SessionSummarizer) bool {
 	return summarizer != nil
 }
 
@@ -41,13 +24,13 @@ func HasSummarizer(summarizer SessionSummarizer) bool {
 // context-aware summary path when available.
 func ShouldSummarize(
 	ctx context.Context,
-	summarizer SessionSummarizer,
+	summarizer summary.SessionSummarizer,
 	sess *session.Session,
 ) bool {
 	if summarizer == nil {
 		return false
 	}
-	if contextual, ok := summarizer.(ContextAwareSummarizer); ok {
+	if contextual, ok := summarizer.(summary.ContextAwareSummarizer); ok {
 		return contextual.ShouldSummarizeWithContext(ctx, sess)
 	}
 	return summarizer.ShouldSummarize(sess)
